refactor(jira): extract wiki markup to its own module; unify content conversion for descriptions and comments
Addresses two review points on the rich-text change:

1. The wiki-markup serializer was a large, self-contained concern living
   in jiraDeployment.ts (which is about Cloud-vs-DC dialect decisions).
   Moved adfToWikiMarkup + its helpers to a dedicated jiraWikiMarkup.ts
   (with jiraWikiMarkup.test.ts), leaving jiraDeployment.ts focused on
   detection/auth/user-ref concerns.

2. Rich text is not description-only — issue comment bodies use the same
   wiki-markup (DC) / ADF (Cloud) grammar. The conversion was duplicated:
   createIssue and updateIssue each carried an identical ~35-line
   description block, and addComment had its own separate inline body
   construction. Unified all three onto one deployment-aware
   toJiraContent(input) helper (Server/DC -> wiki markup string, Cloud ->
   ADF doc, handling TipTap/HTML/plain input and empty values). Comments
   now go through the same pipeline, so they're no longer the one
   rich-text field silently limited to plain text.

Behavior is unchanged for every current caller: descriptions convert
exactly as before (unit assertions pin the wiki-markup output), and
addComment's sole caller (linkToTestCase) passes plain text, which
converts to itself on DC and a single ADF paragraph on Cloud — identical
to the previous hand-built body.

Verified: 131 unit tests green (dialect + wiki-markup + adapter); live
end-to-end against Jira DC 10.3.13 — a formatted TipTap description and
a comment both round-trip correctly through toJiraContent. eslint + tsc
clean.
