+++ testplanit/lib/integrations/adapters/jiraDeployment.ts
@@ -220,171 +220,3 @@ export function mapCustomFieldUserRefs(
   }
   return mapped;
 }
-
-/**
- * Serialize an Atlassian Document Format (ADF) document to Jira Wiki Markup —
- * the markup language Jira Server / Data Center's REST API v2 stores and
- * renders in rich-text fields (issue descriptions). Cloud takes ADF
- * directly; only Server/DC needs this. It is the write-side counterpart to
- * the adapter's ADF→HTML reader, covering the same block and inline set so
- * content round-trips: bold/italic/underline/strike/code/link marks,
- * headings, nested bullet/ordered lists, block quotes, code blocks (with
- * language), horizontal rules, hard breaks, and tables.
- *
- * DC v2 previously received descriptions stripped to bare plain text, which
- * silently discarded every bit of formatting a user entered. It is NOT
- * plain text: a plain string handed to Jira is itself interpreted as wiki
- * markup (`*x*` renders bold), which is exactly why the adapter emits wiki
- * markup here rather than raw text or ADF.
- *
- * Literal text is emitted as-is, not escaped: ADF separates a text node's
- * content from its marks, so "*star*" only renders bold if it actually
- * carries a `strong` mark. Text that itself contains wiki metacharacters is
- * a rare edge Jira will interpret — accepted as a known limitation rather
- * than risking mangled output from aggressive escaping.
- */
-export function adfToWikiMarkup(doc: unknown): string {
-  const content = (doc as { content?: unknown[] } | null)?.content;
-  if (!Array.isArray(content)) return "";
-  // Blocks self-terminate with a blank line; collapse the runs that produces
-  // (and any empty paragraphs) down to a single blank-line separator.
-  return serializeWikiBlocks(content)
-    .replace(/\n{3,}/g, "\n\n")
-    .trim();
-}
-
-function serializeWikiBlocks(nodes: unknown[]): string {
-  return nodes.map((n) => serializeWikiBlock(n)).join("");
-}
-
-function serializeWikiBlock(node: any): string {
-  if (!node || typeof node !== "object") return "";
-  switch (node.type) {
-    case "paragraph":
-      return `${serializeWikiInline(node.content)}\n\n`;
-    case "heading": {
-      const level = Math.min(Math.max(Number(node.attrs?.level) || 1, 1), 6);
-      return `h${level}. ${serializeWikiInline(node.content)}\n\n`;
-    }
-    case "bulletList":
-    case "orderedList":
-      return `${serializeWikiList(node, "")}\n\n`;
-    case "blockquote":
-      return `{quote}\n${serializeWikiBlocks(node.content || []).trim()}\n{quote}\n\n`;
-    case "codeBlock": {
-      const language = node.attrs?.language;
-      const code = (node.content || [])
-        .map((c: any) => (typeof c?.text === "string" ? c.text : ""))
-        .join("");
-      return `{code${language ? `:${language}` : ""}}\n${code}\n{code}\n\n`;
-    }
-    case "rule":
-      return "----\n\n";
-    case "table":
-      return `${serializeWikiTable(node)}\n\n`;
-    default:
-      // Unknown block: recurse into block children, else treat as inline.
-      if (Array.isArray(node.content)) return serializeWikiBlocks(node.content);
-      return serializeWikiInline([node]);
-  }
-}
-
-function serializeWikiInline(content: unknown): string {
-  if (!Array.isArray(content)) return "";
-  return content
-    .map((node: any) => {
-      if (!node || typeof node !== "object") return "";
-      if (node.type === "hardBreak") return "\n";
-      if (node.type === "text") {
-        return applyWikiMarks(node.text || "", node.marks);
-      }
-      if (Array.isArray(node.content)) return serializeWikiInline(node.content);
-      return "";
-    })
-    .join("");
-}
-
-function applyWikiMarks(text: string, marks: unknown): string {
-  if (!Array.isArray(marks) || marks.length === 0) return text;
-  let out = text;
-  // Tolerant of both ADF (strong/em) and raw TipTap (bold/italic) mark names
-  // so the serializer works whether or not the doc was normalized first.
-  for (const mark of marks) {
-    switch ((mark as { type?: string })?.type) {
-      case "strong":
-      case "bold":
-        out = `*${out}*`;
-        break;
-      case "em":
-      case "italic":
-        out = `_${out}_`;
-        break;
-      case "underline":
-        out = `+${out}+`;
-        break;
-      case "strike":
-      case "strikethrough":
-        out = `-${out}-`;
-        break;
-      case "code":
-        out = `{{${out}}}`;
-        break;
-      case "link": {
-        const href = (mark as { attrs?: { href?: string } })?.attrs?.href;
-        out = href ? `[${out}|${href}]` : out;
-        break;
-      }
-    }
-  }
-  return out;
-}
-
-/**
- * Serialize a bullet/ordered list to wiki markup. `prefix` accumulates the
- * ancestor markers so nested and mixed lists render correctly — a bullet
- * under an ordered item becomes `#*`, matching Jira's own nesting syntax,
- * rather than being flattened to a single level.
- */
-function serializeWikiList(listNode: any, prefix: string): string {
-  const marker = listNode.type === "orderedList" ? "#" : "*";
-  const myPrefix = prefix + marker;
-  const lines: string[] = [];
-  for (const item of listNode.content || []) {
-    if (!item || item.type !== "listItem") continue;
-    let line = "";
-    const nested: string[] = [];
-    for (const child of item.content || []) {
-      if (!child || typeof child !== "object") continue;
-      if (child.type === "bulletList" || child.type === "orderedList") {
-        nested.push(serializeWikiList(child, myPrefix));
-      } else if (child.type === "paragraph") {
-        line += serializeWikiInline(child.content);
-      } else {
-        line += serializeWikiInline(child.content || [child]);
-      }
-    }
-    lines.push(`${myPrefix} ${line}`.trimEnd());
-    for (const n of nested) if (n) lines.push(n);
-  }
-  return lines.join("\n");
-}
-
-function serializeWikiTable(node: any): string {
-  const rows: string[] = [];
-  for (const row of node.content || []) {
-    if (!row || row.type !== "tableRow") continue;
-    const cells = (row.content || []).filter(Boolean);
-    const allHeaders =
-      cells.length > 0 &&
-      cells.every((c: any) => c.type === "tableHeader");
-    const rendered = cells.map((c: any) =>
-      serializeWikiBlocks(c.content || [])
-        .replace(/\s+/g, " ")
-        .trim()
-    );
-    rows.push(
-      allHeaders ? `||${rendered.join("||")}||` : `|${rendered.join("|")}|`
-    );
-  }
-  return rows.join("\n");
-}
